Compiler warning about double->uint64_t implicit conversion

+++ util/usage.cc
@@ -135,7 +135,7 @@ template <class Num> uint64_t ParseNum(const std::string &arg) {
   if (after == "%") {
     uint64_t mem = GuessPhysicalMemory();
     UTIL_THROW_IF_ARG(!mem, SizeParseError, (arg), "because % was specified but the physical memory size could not be determined.");
-    return static_cast<double>(value) * static_cast<double>(mem) / 100.0;
+    return static_cast<uint64_t>(static_cast<double>(value) * static_cast<double>(mem) / 100.0);
   }
   
   std::string units("bKMGTPEZY");
